+++ test/reporter-cucumber-html/check-report/passed.test.ts
@@ -35,6 +35,7 @@ test('Scenario: Scenario with attachments', async ({ page }) => {
     'text attachmentsome text', // no space between 'attachment' and 'some'
     'image attachment inline',
     'image attachment as file',
+    'stdoutsome logs',
   ]);
 });