+++ test/reporter-cucumber-html/features/steps.ts
@@ -28,6 +28,11 @@ Then('attach image as file', async ({ $testInfo }) => {
   });
 });
 
+Then('attach stdout', async () => {
+  console.log('some logs'); // eslint-disable-line no-console
+  // don't test console.error b/c it poisons the output
+});
+
 When('open page {string}', async ({ page }, url: string) => {
   await page.goto(url);
 });